// src/api/subscription/api.ts
export interface BillingOptionDto {
  /**
   *
   * @type {string}
   * @memberof BillingOptionDto
   */
  interval?: BillingOptionDtoIntervalEnum;
  /**
   *
   * @type {number}
   * @memberof BillingOptionDto
   */
  price?: number;
  /**
   *
   * @type {string}
   * @memberof BillingOptionDto
   */
  currency?: string;
  /**
   *
   * @type {string}
   * @memberof BillingOptionDto
   */
  priceId?: string;
}

export const BillingOptionDtoIntervalEnum = {
  Monthly: 'monthly',
  Annually: 'annually',
} as const;

export type BillingOptionDtoIntervalEnum =
  (typeof BillingOptionDtoIntervalEnum)[keyof typeof BillingOptionDtoIntervalEnum];
// ...
export interface SubscriptionPlanDto {
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  price?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  currency?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  interval?: string;
  /**
   *
   * @type {string}
   * @memberof SubscriptionPlanDto
   */
  description?: string;
  /**
   *
   * @type {boolean}
   * @memberof SubscriptionPlanDto
   */
  recommended?: boolean;
  /**
   *
   * @type {Array<string>}
   * @memberof SubscriptionPlanDto
   */
  features?: Array<string>;
  /**
   *
   * @type {Array<BillingOptionDto>}
   * @memberof SubscriptionPlanDto
   */
  billingOptions?: Array<BillingOptionDto>;
}